Block/requirement.txt
keras
requests
numpy
matplotlib
tensorflow
Pillow